refactor(cerebrum-api): split debrief/router.ts to satisfy oxlint max-lines
Extracted shared schemas/types into schemas.ts, helpers
(findSessionById, createOrReplacePendingSession) into helpers.ts, and
the procedures into procedures/{read,write,delete}.ts. router.ts now
composes the three procedure groups into debriefRouter.

No behavior change; all existing tests pass unchanged.
